modules/communication-sync/code/product-service-grpc/requirements.txt
grpcio
grpcio-tools
protobuf
sqlalchemy